requirements.txt
autopep8==1.5.7
pycodestyle==2.7.0
black==21.9b0
Flask==2.0.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.25
Jinja2==3.0.1
PyMySQL==1.0.2
python-dotenv==0.19.0
Bootstrap-Flask==1.8.0
email-validator==1.1.3
Flask-WTF==0.15.1
Flask-Bcrypt==0.7.1
WTForms-SQLAlchemy==0.3
marshmallow==3.14.0
